requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.5
colorama==0.4.6
exceptiongroup==1.1.2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
pydantic==1.10.11
pydantic_core==2.3.0
PyJWT==2.7.0
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.0
watchfiles==0.19.0
websockets==11.0.3
